refactor(agent): extract agent assignment logic from execution supervisor
